Add handling SQLiteDatabaseCorruptException
When SQLiteDatabaseCorruptException occurred on execute() or executeForString(),
DefaultDatabaseErrorHandler.onCorrupt() is not to be called.
So, Malformed DB file can't be deleted.
Because, Instead SQLiteDatabaseCorruptException, code is throwing SQLiteException.
On boot time, system can't open DB file and it throws RuntimeException.
This causes infinite reset.
So, I modified to throw SQLiteDatabaseCorruptException.

Change-Id: I8613fb3fd1b203e730d8a5c5df525d6d037c44b5
Signed-off-by: BijiOnta <sengolcaduk32@gmail.com>